ROTO!! (incompatible types: possible lossy conversion from long to int) Cambio de int a long para todos los identificadores de los objetos. Hay un problema al compilar porque en alguna parte ahora se devuelve un long y se utiliza como int, lo cual produce la excepción de tipos incompatibles.

+++ app/src/main/java/com/ubung/tc/ubungmobile/controlador/DescriptionSportActivity.java
@@ -50,7 +50,7 @@ public class DescriptionSportActivity extends ActionBarActivity {
     private void InitLabels() {
 
         Intent t = getIntent();
-        int position = Integer.parseInt(t.getStringExtra(MainUbungActivity.POSITION));
+        long position = Long.parseLong(t.getStringExtra(MainUbungActivity.POSITION));
         String user = t.getStringExtra(MainUbungActivity.USER);
         String deporte = "Basketball";
         String descripcion = "hola";
