+++ app/src/main/java/com/ubung/tc/ubungmobile/modelo/persistencia/entidades/Deporte.java
@@ -5,19 +5,19 @@ package com.ubung.tc.ubungmobile.modelo.persistencia.entidades;
  */
 public class Deporte {
 
-    private int id;
+    private long id;
     private String nombre;
     private String nombreArchivoImagen;
     private String descripcion;
 
-    public Deporte(int id, String nombre, String nombreArchivoImagen, String descripcion) {
+    public Deporte(long id, String nombre, String nombreArchivoImagen, String descripcion) {
         this.id = id;
         this.nombre = nombre;
         this.nombreArchivoImagen = nombreArchivoImagen;
         this.descripcion = descripcion;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
