+++ app/src/main/java/com/ubung/tc/ubungmobile/modelo/persistencia/entidades/Zona.java
@@ -5,19 +5,19 @@ package com.ubung.tc.ubungmobile.modelo.persistencia.entidades;
  */
 public class Zona {
 
-    private int id;
+    private long id;
     private String nombre;
     private double[] latLongZoom;
     private int radio;
 
-    public Zona(int id, String nombre, double[] latLongZoom, int radio) {
+    public Zona(long id, String nombre, double[] latLongZoom, int radio) {
         this.id = id;
         this.nombre = nombre;
         this.latLongZoom = latLongZoom;
         this.radio = radio;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
